# improve performance of ModuleConcatenationPlugin for loop