--- a/lib/optimize/ModuleConcatenationPlugin.js
+++ b/lib/optimize/ModuleConcatenationPlugin.js
@@ -281,6 +281,10 @@ class ModuleConcatenationPlugin {
 							for (const reason of newModule.reasons) {
 								reason.dependency.module = newModule;
 							}
+							// TODO: remove when LTS node version contains fixed v8 version
+							// @see https://github.com/webpack/webpack/pull/6613
+							// Turbofan does not correctly inline for-of loops with polymorphic input arrays.
+							// Work around issue by using a standard for loop and assigning dep.module.reasons
 							for (let i = 0; i < newModule.dependencies.length; i++) {
 								let dep = newModule.dependencies[i];
 								if (dep.module) {