@ooflorent 2018-02-28
Such difference is really odd. When I made these changes and benchmarked them, the results were on par. It looks like a _bug_ in v8. The escape analysis should produce optimal code for this hot path. @bmeurer, can you advice us on this?