@bmeurer 2018-02-28
@psmarshall will take a look when you have a repro. If it's a small fix we might be able to patch this in Node 8 as well.