@ooflorent 2018-02-28
@bmeurer no, it's not monomorphic. `dep` instances may be created from different classes that share the same base class (`Dependency`).