@clydin 2018-02-28
While this appears to be a v8 defect, this is causing a major performance problems with large Angular CLI projects; essentially wiping out and reversing the performance benefits of Webpack 4.  Please considering merging this as soon as possible.