@bmeurer 2018-03-01
Ok, I have a repro for V8. It's because the inlining of `Array.prototype[Symbol.iterator]` and `ArrayIterator.prototype.next` get's confused by polymorphic input arrays. Simple test case to illustrate the problem:

```js
const N = 1e7;

function forOf(a) {
  for (const x of a) {}
}

function test(f, a) {
  let result;
  for (let i = 0; i < N; ++i) result = f(a);
  return result;
}

const a = [1,2,3,4];
const b = ['a', 'b', 'c', 'd'];

test(forOf, a);
for (let i = 0; i < 100; ++i) test(x => x, a);

console.time('mono');
test(forOf, a);
console.timeEnd('mono');

test(forOf, b);

console.time('poly');
test(forOf, a);
console.timeEnd('poly');
```

Filed [v8:7510](crbug.com/v8/7510) upstream.