@bmeurer 2018-03-06
Thanks for the heads-up. The bug ([v8:7510](crbug.com/v8/7510)) was just fixed upstream. Unfortunately just missed 6.6, so it might not make it into Node 10, unless we can float this patch on top.